# packages/contracts/http/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Place HTTP",
    "version": "1.0.0",
    "description": "Place-owned source-only HTTP and browser BFF contract."
  },
  "servers": [],
  "paths": {
...
    "/api/membership-consents/current": {
      "get": {
        "operationId": "getCurrentPlaceMembershipConsentsForBrowser",
        "summary": "getCurrentPlaceMembershipConsentsForBrowser",
        "security": [],
        "responses": {
          "200": {
            "description": "Return current consent versions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentMembershipConsents"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/BrowserMembershipUnavailable"
          }
        }
      }
    },
    "/api/memberships/onboarding": {
      "post": {
        "operationId": "completePlaceBrowserMembershipOnboarding",
        "summary": "completePlaceBrowserMembershipOnboarding",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MembershipOnboardingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Return the existing membership",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipOnboardingResult"
                }
              }
            }
          },
          "201": {
            "description": "Create a non-elevated membership",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipOnboardingResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/OnboardingRequestInvalid"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "409": {
            "$ref": "#/components/responses/MembershipConsentRequired"
          },
          "503": {
            "$ref": "#/components/responses/BrowserMembershipUnavailable"
          }
        }
      }
    },
...
    "/v1/me": {
      "get": {
        "operationId": "getCurrentPlaceMembership",
        "summary": "getCurrentPlaceMembership",
        "security": [
          {
            "placeBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Return the safe current membership",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentMembership"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/AccessDenied"
          }
        }
      }
    },
    "/v1/memberships/onboarding": {
      "post": {
        "operationId": "completePlaceMembershipOnboarding",
        "summary": "completePlaceMembershipOnboarding",
        "security": [
          {
            "placeBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MembershipOnboardingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Return an existing membership",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipOnboardingResult"
                }
              }
            }
          },
          "201": {
            "description": "Create a non-elevated membership",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipOnboardingResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/OnboardingRequestInvalid"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "409": {
            "$ref": "#/components/responses/MembershipConsentRequired"
          },
          "503": {
            "$ref": "#/components/responses/MembershipOnboardingUnavailable"
          }
        }
      }
    },
    "/v1/membership-consents/current": {
      "get": {
        "operationId": "getCurrentPlaceMembershipConsents",
        "summary": "getCurrentPlaceMembershipConsents",
        "security": [],
        "responses": {
          "200": {
            "description": "Return current consent versions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentMembershipConsents"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/MembershipOnboardingUnavailable"
          }
        }
      }
    },
    "/v1/administration/memberships/{membershipId}/authority-role": {
      "parameters": [
        {
          "name": "membershipId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "patch": {
        "operationId": "changePlaceMembershipAuthorityRole",
        "summary": "changePlaceMembershipAuthorityRole",
        "security": [
          {
            "placeBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorityRoleChangeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Return the audited authority mutation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorityRoleChangeResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/AuthorityRequestInvalid"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/AccessDenied"
          },
          "404": {
            "$ref": "#/components/responses/MembershipNotFound"
          },
          "409": {
            "$ref": "#/components/responses/AuthorityChangeRejected"
          },
          "503": {
            "$ref": "#/components/responses/AuthorityChangeUnavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "placeBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "placeBrowserSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-place_session"
      },
      "placeConnector": {
        "type": "http",
        "scheme": "PlaceConnector"
      }
    },
    "schemas": {
...
      "CurrentMembership": {
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "string",
            "const": "place-current-membership.v1"
          },
          "membershipId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "authorityRole": {
            "type": "string",
            "enum": [
              "member",
              "reviewer",
              "administrator",
              "owner"
            ]
          },
          "userGrade": {
            "type": "string",
            "minLength": 1
          },
          "productTier": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "schemaVersion",
          "membershipId",
          "authorityRole",
          "userGrade",
          "productTier"
        ],
        "additionalProperties": false
      },
...
      "MembershipOnboardingRequest": {
        "type": "object",
        "properties": {
          "acceptedConsents": {
            "minItems": 1,
            "maxItems": 32,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "document": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                },
                "version": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                }
              },
              "required": [
                "document",
                "version"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "acceptedConsents"
        ],
        "additionalProperties": false
      },
      "CurrentMembershipConsents": {
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "string",
            "const": "place-membership-consents.v1"
          },
          "consents": {
            "minItems": 1,
            "maxItems": 32,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "document": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                },
                "version": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                }
              },
              "required": [
                "document",
                "version"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "schemaVersion",
          "consents"
        ],
        "additionalProperties": false
      },
      "MembershipOnboardingResult": {
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "string",
            "const": "place-membership-onboarding-result.v1"
          },
          "status": {
            "type": "string",
            "enum": [
              "created",
              "existing"
            ]
          },
          "membershipId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "authorityRole": {
            "type": "string",
            "enum": [
              "member",
              "reviewer",
              "administrator",
              "owner"
            ]
          },
          "userGrade": {
            "type": "string",
            "minLength": 1
          },
          "productTier": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "schemaVersion",
          "status",
          "membershipId",
          "authorityRole",
          "userGrade",
          "productTier"
        ],
        "additionalProperties": false
      },
      "AuthorityRoleChangeRequest": {
        "type": "object",
        "properties": {
          "nextRole": {
            "type": "string",
            "enum": [
              "member",
              "reviewer",
              "administrator"
            ]
          }
        },
        "required": [
          "nextRole"
        ],
        "additionalProperties": false
      },
      "AuthorityRoleChangeResult": {
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "string",
            "const": "place-authority-role-change-result.v1"
          },
          "status": {
            "type": "string",
            "enum": [
              "changed",
              "unchanged"
            ]
          },
          "membershipId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "previousRole": {
            "type": "string",
            "enum": [
              "member",
              "reviewer",
              "administrator",
              "owner"
            ]
          },
          "authorityRole": {
            "type": "string",
            "enum": [
              "member",
              "reviewer",
              "administrator",
              "owner"
            ]
          }
        },
        "required": [
          "schemaVersion",
          "status",
          "membershipId",
          "authorityRole"
        ],
        "additionalProperties": false
      },
...
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "minLength": 1
          },
          "title": {
            "type": "string",
            "minLength": 1
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "code": {
            "type": "string",
            "minLength": 1
          },
          "retryable": {
            "type": "boolean"
          },
          "correlationRef": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "retryable",
          "correlationRef"
        ],
        "additionalProperties": false
      }
    },
    "responses": {
...
      "AuthenticationRequired": {
        "description": "The bearer token is missing or invalid",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "AccessDenied": {
        "description": "The principal lacks Place access",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "BrowserMembershipUnavailable": {
        "description": "Browser membership is inactive or unavailable",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "OnboardingRequestInvalid": {
        "description": "The onboarding request is invalid",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "MembershipConsentRequired": {
        "description": "Current Place consent is required",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "MembershipOnboardingUnavailable": {
        "description": "Membership onboarding is unavailable",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "AuthorityRequestInvalid": {
        "description": "The authority-role request is invalid",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "MembershipNotFound": {
        "description": "The target membership does not exist",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "AuthorityChangeRejected": {
        "description": "The authority-role change was rejected",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "AuthorityChangeUnavailable": {
        "description": "Authority-role management is unavailable",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    }
  }
}